Delete only files

+++ dele.go
@@ -11,6 +11,10 @@ func (c *Connection) dele(args []string) {
 		return
 	}
 	filename := filepath.Join(c.curDir(), filepath.Clean(args[0]))
+	if fileinfo, err := os.Stat(filename); err == nil && fileinfo.IsDir() {
+		c.writeout("450 Requested file action not taken.")
+		return
+	}
 	err := os.Remove(filename)
 	if err != nil {
 		c.writeout("550 Requested action not taken. File unavailable.")
